Add TypeScript type definitions (#2)
TypeScript will refuse to build a project that relies on `eth-query`,
as it lacks type definitions. We have previously gotten around this in
the extension and in `core` by supplying those type definitions directly
in the project itself, but this is causing issues especially for the
Snaps team, who now wish to use `@metamask/controller-utils`, which
names `eth-query` as a dependency.

This commit copies the type definitions from the `core` monorepo into
this repo. They were previously copied from the extension, so they have
been battle-tested pretty well now.